+++ projects/pip-webui2-controls/src/lib/check-item/check-item.component.spec.ts
@@ -1,25 +1,30 @@
-import { TestBed, inject, async } from '@angular/core/testing';
+import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
+import { TestBed, ComponentFixture } from '@angular/core/testing';
 
 import { PipCheckItemComponent } from './check-item.component';
 
-describe('a pip-ref-item component', () => {
-	let component: PipCheckItemComponent;
+describe('a pip-check-item component', () => {
+    let component: PipCheckItemComponent;
+    let fixture: ComponentFixture<PipCheckItemComponent>;
 
-	// register all needed dependencies
-	beforeEach(() => {
-		TestBed.configureTestingModule({
-			providers: [
-				PipCheckItemComponent
-			]
-		});
-	});
+    // register all needed dependencies
+    beforeEach(() => {
+        TestBed.configureTestingModule({
+            declarations: [
+                PipCheckItemComponent
+            ],
+            schemas: [CUSTOM_ELEMENTS_SCHEMA]
+        });
+        fixture = TestBed.createComponent(PipCheckItemComponent);
+        component = fixture.componentInstance;
+    });
 
-	// instantiation through framework injection
-	beforeEach(inject([PipCheckItemComponent], (PipCheckItemComponent) => {
-		component = PipCheckItemComponent;
-	}));
+    // instantiation through framework injection
+    // beforeEach(inject([PipCheckItemComponent], (CheckItemComponent) => {
+    //     component = CheckItemComponent;
+    // }));
 
-	it('should have an instance', () => {
-		expect(component).toBeDefined();
-	});
+    it('should have an instance', () => {
+        expect(component).toBeDefined();
+    });
 });