+++ projects/pip-webui2-controls/src/lib/empty-state/empty-state.component.spec.ts
@@ -1,25 +1,23 @@
-import { TestBed, inject } from '@angular/core/testing';
+import { TestBed, ComponentFixture } from '@angular/core/testing';
 
 import { PipEmptyStateComponent } from './empty-state.component';
 
-describe('a empty-state component', () => {
-	let component: PipEmptyStateComponent;
+describe('a pip-empty-state component', () => {
+    let component: PipEmptyStateComponent;
+    let fixture: ComponentFixture<PipEmptyStateComponent>;
 
-	// register all needed dependencies
-	beforeEach(() => {
-		TestBed.configureTestingModule({
-			providers: [
-				PipEmptyStateComponent
-			]
-		});
-	});
+    // register all needed dependencies
+    beforeEach(() => {
+        TestBed.configureTestingModule({
+            declarations: [
+                PipEmptyStateComponent
+            ]
+        });
+        fixture = TestBed.createComponent(PipEmptyStateComponent);
+        component = fixture.componentInstance;
+    });
 
-	// instantiation through framework injection
-	beforeEach(inject([PipEmptyStateComponent], (PipEmptyStateComponent) => {
-		component = PipEmptyStateComponent;
-	}));
-
-	it('should have an instance', () => {
-		expect(component).toBeDefined();
-	});
+    it('should have an instance', () => {
+        expect(component).toBeDefined();
+    });
 });