+++ projects/pip-webui2-controls/src/lib/ref-item/ref-item.component.spec.ts
@@ -1,25 +1,25 @@
-import { TestBed, inject } from '@angular/core/testing';
+import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
+import { TestBed, ComponentFixture } from '@angular/core/testing';
 
 import { PipRefItemComponent } from './ref-item.component';
 
 describe('a pip-ref-item component', () => {
-	let component: PipRefItemComponent;
+    let component: PipRefItemComponent;
+    let fixture: ComponentFixture<PipRefItemComponent>;
 
-	// register all needed dependencies
-	beforeEach(() => {
-		TestBed.configureTestingModule({
-			providers: [
-				PipRefItemComponent
-			]
-		});
-	});
+    // register all needed dependencies
+    beforeEach(() => {
+        TestBed.configureTestingModule({
+            declarations: [
+                PipRefItemComponent
+            ],
+            schemas: [CUSTOM_ELEMENTS_SCHEMA]
+        });
+        fixture = TestBed.createComponent(PipRefItemComponent);
+        component = fixture.componentInstance;
+    });
 
-	// instantiation through framework injection
-	beforeEach(inject([PipRefItemComponent], (PipRefItemComponent) => {
-		component = PipRefItemComponent;
-	}));
-
-	it('should have an instance', () => {
-		expect(component).toBeDefined();
-	});
+    it('should have an instance', () => {
+        expect(component).toBeDefined();
+    });
 });