+++ src/app/app.component.spec.ts
@@ -1,27 +1,57 @@
 import { TestBed, async } from '@angular/core/testing';
+import { NoopAnimationsModule } from '@angular/platform-browser/animations';
+import { FlexLayoutModule } from '@angular/flex-layout';
+import { MatToolbarModule, MatSelectModule, MatSidenavModule, MatIconModule, MatButtonModule, MatMenuModule } from '@angular/material';
+import { TranslateModule } from '@ngx-translate/core';
+import { PipThemesModule } from 'pip-webui2-themes';
+
 import { AppComponent } from './app.component';
+import { AppRoutingModule } from './app-routing.module';
+import { CheckListExampleModule } from './check-list-example/check-list-example.module';
+import { ColorPickerExampleModule } from './color-picker-example/color-picker-example.module';
+import { EmptyStateExampleModule } from './empty-state-example/empty-state-example.module';
+import { JumbotronExampleModule } from './jumbotron-example/jumbotron-example.module';
+import { ExamplesListModule } from './examples-list/examples-list.module';
+import { PartsExampleModule } from './parts-example/parts-example.module';
+import { RefListExampleModule } from './ref-list-example/ref-list-example.module';
+import { SearchInputExampleModule } from './search-input-example/search-input-example.module';
+import { SliderExampleModule } from './slider-example/slider-example.module';
+
 describe('AppComponent', () => {
   beforeEach(async(() => {
     TestBed.configureTestingModule({
       declarations: [
         AppComponent
       ],
+      imports: [
+        NoopAnimationsModule,
+        FlexLayoutModule,
+        MatToolbarModule,
+        MatSelectModule,
+        MatSidenavModule,
+        MatIconModule,
+        MatButtonModule,
+        MatMenuModule,
+        TranslateModule.forRoot(),
+
+        PipThemesModule,
+
+        AppRoutingModule,
+        CheckListExampleModule,
+        ColorPickerExampleModule,
+        EmptyStateExampleModule,
+        JumbotronExampleModule,
+        ExamplesListModule,
+        PartsExampleModule,
+        RefListExampleModule,
+        SearchInputExampleModule,
+        SliderExampleModule
+      ]
     }).compileComponents();
   }));
   it('should create the app', async(() => {
     const fixture = TestBed.createComponent(AppComponent);
     const app = fixture.debugElement.componentInstance;
     expect(app).toBeTruthy();
-  }));
-  it(`should have as title 'pip-webui2-dates-app'`, async(() => {
-    const fixture = TestBed.createComponent(AppComponent);
-    const app = fixture.debugElement.componentInstance;
-    expect(app.title).toEqual('pip-webui2-dates-app');
-  }));
-  it('should render title in a h1 tag', async(() => {
-    const fixture = TestBed.createComponent(AppComponent);
-    fixture.detectChanges();
-    const compiled = fixture.debugElement.nativeElement;
-    expect(compiled.querySelector('h1').textContent).toContain('Welcome to pip-webui2-dates-app!');
   }));
 });