+++ src/app/app.module.ts
@@ -1,7 +1,7 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { NgModule } from '@angular/core';
-import { FlexLayoutModule, BREAKPOINTS, DEFAULT_BREAKPOINTS } from '@angular/flex-layout';
+import { FlexLayoutModule } from '@angular/flex-layout';
 import { MatToolbarModule, MatSelectModule, MatSidenavModule, MatIconModule, MatButtonModule, MatMenuModule } from '@angular/material';
 import { TranslateModule } from '@ngx-translate/core';
 import { PipThemesModule } from 'pip-webui2-themes';
@@ -17,12 +17,6 @@ import { PartsExampleModule } from './parts-example/parts-example.module';
 import { RefListExampleModule } from './ref-list-example/ref-list-example.module';
 import { SearchInputExampleModule } from './search-input-example/search-input-example.module';
 import { SliderExampleModule } from './slider-example/slider-example.module';
-
-export const CustomBreakPointsProvider = {
-  provide: BREAKPOINTS,
-  useValue: [...DEFAULT_BREAKPOINTS],
-  multi: true
-};
 
 @NgModule({
   declarations: [
@@ -53,7 +47,7 @@ export const CustomBreakPointsProvider = {
     SearchInputExampleModule,
     SliderExampleModule
   ],
-  providers: [CustomBreakPointsProvider],
+  providers: [],
   bootstrap: [AppComponent]
 })
 export class AppModule { }