+++ src/app/check-list-example/check-list-example.component.spec.ts
@@ -1,4 +1,8 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { CommonModule } from '@angular/common';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+import { MatListModule } from '@angular/material';
+import { PipCheckItemModule } from 'pip-webui2-controls';
 
 import { CheckListExampleComponent } from './check-list-example.component';
 
@@ -8,9 +12,17 @@ describe('CheckListExampleComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ CheckListExampleComponent ]
+      declarations: [CheckListExampleComponent],
+      imports: [
+        CommonModule,
+        FormsModule,
+        ReactiveFormsModule,
+        MatListModule,
+
+        PipCheckItemModule
+      ],
     })
-    .compileComponents();
+      .compileComponents();
   }));
 
   beforeEach(() => {