+++ src/app/color-picker-example/color-picker-example.component.spec.ts
@@ -1,4 +1,6 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { CommonModule } from '@angular/common';
+import { PipColorPickerModule } from 'pip-webui2-controls';
 
 import { ColorPickerExampleComponent } from './color-picker-example.component';
 
@@ -8,9 +10,14 @@ describe('ColorPickerExampleComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ ColorPickerExampleComponent ]
+      declarations: [ColorPickerExampleComponent],
+      imports: [
+        CommonModule,
+
+        PipColorPickerModule
+      ],
     })
-    .compileComponents();
+      .compileComponents();
   }));
 
   beforeEach(() => {