+++ src/app/empty-state-example/empty-state-example.component.spec.ts
@@ -1,4 +1,7 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { CommonModule } from '@angular/common';
+import { TranslateModule } from '@ngx-translate/core';
+import { PipEmptyStateModule } from 'pip-webui2-controls';
 
 import { EmptyStateExampleComponent } from './empty-state-example.component';
 
@@ -8,9 +11,15 @@ describe('EmptyStateExampleComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ EmptyStateExampleComponent ]
+      declarations: [EmptyStateExampleComponent],
+      imports: [
+        CommonModule,
+        TranslateModule.forRoot(),
+
+        PipEmptyStateModule
+      ],
     })
-    .compileComponents();
+      .compileComponents();
   }));
 
   beforeEach(() => {