+++ src/app/parts-example/parts-example.component.spec.ts
@@ -1,5 +1,11 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { CommonModule } from '@angular/common';
+import { FlexLayoutModule } from '@angular/flex-layout';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+import { MatCheckboxModule, MatCardModule, MatToolbarModule, MatButtonModule, MatIconModule } from '@angular/material';
+import { PipPartModule } from 'pip-webui2-controls';
 
+import { PipPartDirective } from './parts-name-example.directive';
 import { PartsExampleComponent } from './parts-example.component';
 
 describe('PartsExampleComponent', () => {
@@ -8,9 +14,22 @@ describe('PartsExampleComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ PartsExampleComponent ]
+      declarations: [PartsExampleComponent, PipPartDirective],
+      imports: [
+        CommonModule,
+        FlexLayoutModule,
+        FormsModule,
+        ReactiveFormsModule,
+        MatCheckboxModule,
+        MatCardModule,
+        MatToolbarModule,
+        MatButtonModule,
+        MatIconModule,
+
+        PipPartModule
+      ],
     })
-    .compileComponents();
+      .compileComponents();
   }));
 
   beforeEach(() => {