+++ src/app/parts-example/parts-name-example.directive.spec.ts
@@ -1,8 +1,0 @@
-import { PipPartDirective } from './parts-name-example.directive';
-
-describe('PartsExampleDirective', () => {
-  it('should create an instance', () => {
-    // const directive = new PipPartDirective();
-    // expect(directive).toBeTruthy();
-  });
-});