+++ src/app/slider-example/slider-example.component.spec.ts
@@ -1,4 +1,9 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { CommonModule } from '@angular/common';
+import { FlexLayoutModule } from '@angular/flex-layout';
+import { NoopAnimationsModule } from '@angular/platform-browser/animations';
+import { MatTabsModule } from '@angular/material';
+import { PipSliderModule } from 'pip-webui2-controls';
 
 import { SliderExampleComponent } from './slider-example.component';
 
@@ -8,9 +13,17 @@ describe('SliderExampleComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ SliderExampleComponent ]
+      declarations: [SliderExampleComponent],
+      imports: [
+        NoopAnimationsModule,
+        CommonModule,
+        FlexLayoutModule,
+        MatTabsModule,
+
+        PipSliderModule
+      ],
     })
-    .compileComponents();
+      .compileComponents();
   }));
 
   beforeEach(() => {