+++ CHANGELOG.md
@@ -59,3 +59,25 @@ No breaking changes in this version
 
 ### Bug Fixes
 * Search field styles improvement
+
+## <a name="1.1.0"></a> 1.1.0 (2018-11-22)
+
+Update to stable Angular 7 and fix the tests.
+
+### Features
+* **check-item**: check item
+* **color-picker**: color picker
+* **empty-state**: custom empty state
+* **jumbotron**: visual container
+* **part**: part of anpther components
+* **ref-item**: ref item
+* **search-input**: search input
+* **slider**: slider
+
+### Breaking Changes
+No breaking changes in this version
+
+### Bug Fixes
+* Library tests works now. Run `npm run test:lib` to check it out;
+* Example tests works now. Run `npm run test` to check it out;
+* E2E tests works too. Run `npm run e2e` to check it out.